+++ telegram/utils/helpers.py
@@ -18,9 +18,7 @@
 # along with this program.  If not, see [http://www.gnu.org/licenses/].
 """ This module contains helper functions """
 
-import copy
 import re
-from telegram import MessageEntity
 
 
 def extract_chat_and_user(update):
@@ -104,100 +102,6 @@ def extract_entities(update):
         return update.callback_query.message.parse_entities()
     else:
         raise ValueError("No message object found in update, therefore no entities available.")
-
-
-def parse_markdown_from_update(update):
-    """
-    Creates a markdown-formatted string from the markup entities found in the supplied update.
-
-    Args:
-        update (:class:`telegram.Update`): The update presumably containing entities
-
-    Returns:
-        str
-
-    Raises:
-        ValueError: If no entities were found in the update
-
-    """
-    entities = copy.deepcopy(extract_entities(update))
-    text = extract_message_text(update)
-
-    for e in entities:
-        inserted = None
-        pos = e.offset
-        length = e.length
-
-        if e.type == MessageEntity.TEXT_LINK:
-            text = "{}[{}]({}){}".format(text[:pos], text[pos:pos + length], e.url,
-                                         text[pos + length:])
-            inserted = 4 + len(e.url)
-        elif e.type == MessageEntity.BOLD:
-            text = text[:pos] + '*' + text[pos:pos + length] + '*' + text[pos + length:]
-            inserted = 2
-        elif e.type == MessageEntity.ITALIC:
-            text = text[:pos] + '_' + text[pos:pos + length] + '_' + text[pos + length:]
-            inserted = 2
-        elif e.type == MessageEntity.CODE:
-            text = text[:pos] + '`' + text[pos:pos + length] + '`' + text[pos + length:]
-            inserted = 2
-        elif e.type == MessageEntity.PRE:
-            text = text[:pos] + '```\n' + text[pos:pos + length] + '\n```' + text[pos + length:]
-            inserted = 6
-
-        # update offsets of all entities to the right
-        if inserted:
-            for other in entities:
-                if other.offset > pos:
-                    other.offset += inserted
-    return text
-
-
-def parse_html_from_update(update):
-    """
-    Creates an html-formatted string from the markup entities found in the supplied update.
-
-    Args:
-        update (:class:`telegram.Update`): The update presumably containing entities
-
-    Returns:
-        str
-
-    Raises:
-        ValueError: if no entities were found in the update
-
-    """
-    entities = copy.deepcopy(extract_entities(update))
-    text = extract_message_text(update)
-
-    for e in entities:
-        inserted = None
-        pos = e.offset
-        length = e.length
-
-        if e.type == MessageEntity.TEXT_LINK:
-            text = '{}<a href="{}">{}</a>{}'.format(text[:pos], e.url, text[pos:pos + length],
-                                                    text[pos + length:])
-            inserted = 15 + len(e.url)
-        elif e.type == MessageEntity.BOLD:
-            text = text[:pos] + '<b>' + text[pos:pos + length] + '</b>' + text[pos + length:]
-            inserted = 7
-        elif e.type == MessageEntity.ITALIC:
-            text = text[:pos] + '<i>' + text[pos:pos + length] + '</i>' + text[pos + length:]
-            inserted = 7
-        elif e.type == MessageEntity.CODE:
-            text = text[:pos] + '<code>' + text[pos:pos + length] + '</code>' + text[pos + length:]
-            inserted = 13
-        elif e.type == MessageEntity.PRE:
-            text = text[:pos] + '<pre>' + text[pos:pos + length] + '</pre>' + text[pos + length:]
-            inserted = 11
-
-        # update offsets of all entities to the right
-        if inserted:
-            for other in entities:
-                if other.offset > pos:
-                    other.offset += inserted
-    return text
 
 
 def escape_markdown(text):
