+++ tests/test_message.py
@@ -33,6 +33,45 @@ from tests.base import BaseTest
 class MessageTest(BaseTest, unittest.TestCase):
     """This object represents Tests for Telegram MessageTest."""
 
+    def setUp(self):
+        self.test_entities = [
+            {
+                'length': 4,
+                'offset': 9,
+                'type': 'bold'
+            },
+            {
+                'length': 6,
+                'offset': 15,
+                'type': 'italic'
+            },
+            {
+                'length': 4,
+                'offset': 23,
+                'type': 'code'
+            },
+            {
+                'length': 5,
+                'offset': 29,
+                'type': 'text_link',
+                'url': 'http://github.com/'
+            },
+            {
+                'length': 3,
+                'offset': 39,
+                'type': 'pre'
+            },
+        ]
+
+        self.test_text = 'Test for bold, italic, code, links and pre.'
+        self.test_message = telegram.Message(
+            message_id=1,
+            from_user=None,
+            date=None,
+            chat=None,
+            text=self.test_text,
+            entities=[telegram.MessageEntity(**e) for e in self.test_entities])
+
     def test_parse_entity(self):
         text = (b'\\U0001f469\\u200d\\U0001f469\\u200d\\U0001f467'
                 b'\\u200d\\U0001f467\\U0001f431http://google.com').decode('unicode-escape')
@@ -58,6 +97,17 @@ class MessageTest(BaseTest, unittest.TestCase):
         self.assertDictEqual(message.parse_entities(),
                              {entity: 'http://google.com',
                               entity_2: 'h'})
+
+    def test_text_html(self):
+        test_html_string = 'Test for <b>bold</b>, <i>italic</i>, <code>code</code>, ' \
+                           '<a href="http://github.com/">links</a> and <pre>pre</pre>.'
+        text_html = self.test_message.text_html
+        self.assertEquals(test_html_string, text_html)
+
+    def test_text_markdown(self):
+        test_md_string = 'Test for *bold*, _italic_, `code`, [links](http://github.com/) and ```\npre\n```.'
+        text_markdown = self.test_message.text_markdown
+        self.assertEquals(test_md_string, text_markdown)
 
     @flaky(3, 1)
     def test_reply_text(self):
